+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 ### Features
 - Add `--mapping-file <mapping-file>` option to specify a mapping file. This allows you to map the names of the files in the bundle to the names of the files in the previous bundle.
+- Add UI to pair/unpair files and to save the current mapping into a text file which can be loaded with `--mapping-file` or `File | Open File Mappings...`.
 
 ## 2.1.0 (13 Feb 2022)